requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==4.0.1
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
types-passlib==1.7.7.20250602
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
